cli: split application logic CLI from admin
* While including admin commands to REANA Server (now `reana-admin`)
  we run into the problem of mixing responsibilities, this commit
  intends to fix that.
