+++ reana_server/cli.py
@@ -6,314 +6,14 @@
 # REANA is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""REANA Server command line tool.
+"""REANA Server command line tool."""
 
-The REANA Server offers a command line tool for user account operations. An
-administrator can create new user accounts, retrieve users based on given
-filters and export/import all users.
-
-Retrieving administrator token
-------------------------------
-
-By default the server docker image is set to create the administrator account
-on startup, which has the email ``info@reana.io`` and a token which is
-generated using the ``secrets`` python library.
-
-To retrieve the administrator token you can use the `reana-dev` CLI:
-
-.. code-block:: bash
-
-    $ reana-dev setup-environment
-
-or enter the database pod:
-
-.. code-block:: bash
-
-    $ kubectl exec -ti <db-pod-name> /bin/bash
-
-access the Postgresql database:
-
-.. code-block:: bash
-
-    $ psql -U reana
-
-and get all users:
-
-.. code-block:: sql
-
-    > SELECT * FROM user_;
-
-Managing users
---------------
-
-With the administrator access token, new user creation is allowed with:
-
-.. code-block :: bash
-
-    $ flask reana-users create -e=<email> --admin-access-token=<token>
-
-Similarly, to retrieve information for all users:
-
-.. code-block :: bash
-
-    $ flask reana-users get --admin-access-token=<token>
-
-To export all users in current REANA cluster you can use:
-
-.. code-block :: bash
-
-    # users will be exported in CSV format
-    $ flask reana-users export --admin-access-token=<token> > users.csv
-
-And import them in a new cluster as follows:
-
-.. code-block :: bash
-
-    $ flask reana-users import --file users.csv --admin-access-token=<token>
-"""
 import logging
-import os
-import secrets
-import sys
-import traceback
 
 import click
-import tablib
-from flask.cli import with_appcontext
 from reana_commons.config import REANA_LOG_FORMAT, REANA_LOG_LEVEL
-from reana_commons.utils import click_table_printer
-from reana_db.database import Session, init_db
-from reana_db.models import User
 
-from reana_server import config
 from reana_server.scheduler import WorkflowExecutionScheduler
-from reana_server.utils import (_create_user, _export_users, _get_users,
-                                _import_users, create_user_workspace)
-
-
-@click.group()
-def reana_db():
-    """Database management commands."""
-
-
-@reana_db.command('init')
-def db_init():
-    """Initialise database."""
-    try:
-        init_db()
-        click.echo(click.style('DB Created.', fg='green'))
-    except Exception as e:
-        click.echo('Something went wrong: {0}'.format(e))
-        sys.exit(1)
-
-
-@click.group(
-    help='All interaction related to user management on REANA cloud.')
-@click.pass_context
-def reana_users(ctx):
-    """Top level wrapper for user related interaction."""
-    logging.debug(ctx.info_name)
-
-
-@reana_users.command('create_default')
-@click.argument('email')
-@click.option('-i', '--id', 'id_',
-              default=config.ADMIN_USER_ID)
-@with_appcontext
-def users_create_default(email, id_):
-    """Create default user.
-
-    This user has the administrator role
-    and can retrieve other user information as well as create
-    new users.
-    """
-    user_characteristics = {"id_": id_,
-                            "email": email,
-                            }
-    try:
-        user = User.query.filter_by(**user_characteristics).first()
-        if not user:
-            user_characteristics['access_token'] = secrets.token_urlsafe()
-            user = User(**user_characteristics)
-            create_user_workspace(user.get_user_workspace())
-            Session.add(user)
-            Session.commit()
-            click.echo('Created 1st user with access_token: {}'.
-                       format(user_characteristics['access_token']))
-    except Exception as e:
-        click.echo('Something went wrong: {0}'.format(e))
-        sys.exit(1)
-
-
-@reana_users.command(
-    'get',
-    help='Get information of users matching search criteria.')
-@click.option(
-    '--id',
-    help='The id of the user.')
-@click.option(
-    '-e',
-    '--email',
-    help='The email of the user.')
-@click.option(
-    '--user-access-token',
-    help='The access token of the user.')
-@click.option(
-    '--admin-access-token',
-    default=os.environ.get('REANA_ADMIN_ACCESS_TOKEN', None),
-    help='The access token of an administrator.')
-@click.option(
-    '--json',
-    'output_format',
-    flag_value='json',
-    default=None,
-    help='Get output in JSON format.')
-@click.pass_context
-def get_users(ctx, id, email, user_access_token, admin_access_token,
-              output_format):
-    """Return user information. Requires the token of an administrator."""
-    try:
-        response = _get_users(id, email, user_access_token, admin_access_token)
-        headers = ['id', 'email', 'access_token', 'access_token_status']
-        data = []
-        for user in response:
-            data.append((str(user.id_), user.email, str(user.access_token),
-                         str(user.access_token_status)))
-        if output_format:
-            tablib_data = tablib.Dataset()
-            tablib_data.headers = headers
-            for row in data:
-                tablib_data.append(row)
-
-            click.echo(tablib_data.export(output_format))
-        else:
-            click_table_printer(headers, [], data)
-
-    except Exception as e:
-        logging.debug(traceback.format_exc())
-        logging.debug(str(e))
-        click.echo(
-            click.style('User could not be retrieved: \n{}'
-                        .format(str(e)), fg='red'),
-            err=True)
-
-
-@reana_users.command(
-    'create',
-    help='Create a new user.')
-@click.option(
-    '-e',
-    '--email',
-    help='The email of the user.')
-@click.option(
-    '--user-access-token',
-    help='The access token of the user.')
-@click.option(
-    '--admin-access-token',
-    default=os.environ.get('REANA_ADMIN_ACCESS_TOKEN', None),
-    help='The access token of an administrator.')
-@click.pass_context
-def create_user(ctx, email, user_access_token, admin_access_token):
-    """Create a new user. Requires the token of an administrator."""
-    try:
-        response = _create_user(email, user_access_token, admin_access_token)
-        headers = ['id', 'email', 'access_token']
-        data = [(str(response.id_), response.email, response.access_token)]
-        click.echo(
-            click.style('User was successfully created.', fg='green'))
-        click_table_printer(headers, [], data)
-
-    except Exception as e:
-        logging.debug(traceback.format_exc())
-        logging.debug(str(e))
-        click.echo(
-            click.style('User could not be created: \n{}'
-                        .format(str(e)), fg='red'),
-            err=True)
-
-
-@reana_users.command('export')
-@click.option(
-    '--admin-access-token',
-    default=os.environ.get('REANA_ADMIN_ACCESS_TOKEN', None),
-    help='The access token of an administrator.')
-@click.pass_context
-def export_users(ctx, admin_access_token):
-    """Export all users in current REANA cluster."""
-    try:
-        csv_file = _export_users(admin_access_token)
-        click.echo(csv_file.getvalue(), nl=False)
-    except Exception as e:
-        click.secho(
-            'Something went wrong while importing users:\n{}'.format(e),
-            fg='red', err=True)
-
-
-@reana_users.command('import')
-@click.option(
-    '--admin-access-token',
-    default=os.environ.get('REANA_ADMIN_ACCESS_TOKEN', None),
-    help='The access token of an administrator.')
-@click.option(
-    '-f',
-    '--file',
-    'file_',
-    help='A CSV file containing a list of REANA users.',
-    type=click.File())
-@click.pass_context
-def import_users(ctx, admin_access_token, file_):
-    """Import users from file."""
-    try:
-        _import_users(admin_access_token, file_)
-        click.secho('Users successfully imported.', fg='green')
-    except Exception as e:
-        click.secho(
-            'Something went wrong while importing users:\n{}'.format(e),
-            fg='red', err=True)
-
-
-@reana_users.group(
-    'token',
-    help='All interaction related to user tokens on REANA cloud.')
-def reana_users_token():
-    """Manage REANA user tokens."""
-
-
-@reana_users_token.command('grant')
-@click.option(
-    '--id',
-    help='The id of the user.')
-@click.option(
-    '--admin-access-token',
-    default=os.environ.get('REANA_ADMIN_ACCESS_TOKEN', None),
-    help='The access token of an administrator.')
-def token_grant(id, admin_access_token):
-    """."""
-    user_id = 'dc4aef21-c324-4fd5-8480-c724de1f0c06'
-    user_email = 'john.doe@google.com'
-    user_granted_token = 'c0fa47fa00ae4013a13fd7n'
-    # create a token in UserToken table and activate it
-    # send notification to user by email
-    click.secho(f'Token for user {user_id} ({user_email}) granted.\n'
-                f'\nToken: {user_granted_token}', fg='green')
-
-
-@reana_users_token.command('revoke')
-@click.option(
-    '--id',
-    help='The id of the user.')
-@click.option(
-    '--admin-access-token',
-    default=os.environ.get('REANA_ADMIN_ACCESS_TOKEN', None),
-    help='The access token of an administrator.')
-def token_revoke(id, admin_access_token):
-    """."""
-    user_id = 'dc4aef21-c324-4fd5-8480-c724de1f0c06'
-    user_email = 'john.doe@google.com'
-    # deactivate REANA token in UserToken table
-    # send notification to user by email
-    click.secho(f'User\'s {user_id} ({user_email}) token successfully revoked',
-                fg='green')
 
 
 @click.command('start-scheduler')
